refactor: break core<->runner type cycles ahead of runner extraction
Stage 1 of the wholesale homeboy-runner extraction: sever the last core->runner
TYPE cycles so the runner subsystem can move to its own crate without a
core <-> homeboy-runner dependency cycle.

The two core files that named runner types (outside the upgrade/runners
cluster) are dev_run and run_outcome_envelope:

- Moved the plain-data types dev_run needs into homeboy-runner-contract:
  RunnerLifecycleOwner (+ as_str), ByteFileCounts, RunnerWorkspaceLease,
  RunnerWorkspaceCurrentSummary. Re-exported from their old runner paths so
  runner-internal call sites resolve unchanged; dev_run now names them via the
  contract.
- run_outcome_envelope::add_handoff took a runner-owned &RunnerHandoff but only
  read four primitive fields. Changed it to take those fields directly; its one
  caller (runner/worker/result.rs) extracts them. RunnerHandoff (and its
  RunnerJob graph) now stays entirely in runner.

After this, the only remaining non-upgrade core->runner reference is the Runner
ConfigEntity registration in config.rs, which the register_config_entity
keystone handles when runner becomes its own crate. Next stages: invert the
upgrade/helpers -> upgrade/runners boundary + move upgrade/runners with runner,
then the wholesale git mv.

Verified: cargo check clean for -p homeboy-runner-contract, -p homeboy-core
--lib/--tests, -p homeboy-cli --lib.
